[react-interaction] Refactor a11y components more
Format

Fix flow

fix
